+++ common/src/main/java/com/coolerpromc/resourcestrees/ResourcesTreesClient.java
@@ -6,6 +6,6 @@ import net.minecraft.resources.Identifier;
 import java.util.HashMap;
 import java.util.Map;
 
-public class CommonClientClass {
+public class ResourcesTreesClient {
     public static final Map<Identifier, ClientItem> CLIENT_ITEM_MAP = new HashMap<>();
 }