+++ common/src/main/java/com/coolerpromc/resourcestrees/mixin/ModelBakeryMixin.java
@@ -1,6 +1,6 @@
 package com.coolerpromc.resourcestrees.mixin;
 
-import com.coolerpromc.resourcestrees.CommonClientClass;
+import com.coolerpromc.resourcestrees.ResourcesTreesClient;
 import net.minecraft.client.model.geom.EntityModelSet;
 import net.minecraft.client.renderer.PlayerSkinRenderCache;
 import net.minecraft.client.renderer.block.dispatch.BlockStateModel;
@@ -24,6 +24,6 @@ public class ModelBakeryMixin {
             method = "<init>(Lnet/minecraft/client/model/geom/EntityModelSet;Lnet/minecraft/client/resources/model/sprite/SpriteGetter;Lnet/minecraft/client/renderer/PlayerSkinRenderCache;Ljava/util/Map;Ljava/util/Map;Ljava/util/Map;Lnet/minecraft/client/resources/model/ResolvedModel;)V"
     )
     public void init(EntityModelSet entityModelSet, SpriteGetter sprites, PlayerSkinRenderCache playerSkinRenderCache, Map<BlockState, BlockStateModel.UnbakedRoot> unbakedBlockStateModels, Map<Identifier, ClientItem> clientInfos, Map<Identifier, ResolvedModel> resolvedModels, ResolvedModel missingModel, CallbackInfo ci) {
-        clientInfos.putAll(CommonClientClass.CLIENT_ITEM_MAP);
+        clientInfos.putAll(ResourcesTreesClient.CLIENT_ITEM_MAP);
     }
 }