+++ forge/src/main/java/com/coolerpromc/resourcestrees/ForgeResourcesTrees.java
@@ -1,5 +1,7 @@
 package com.coolerpromc.resourcestrees;
 
+import com.coolerpromc.resourcestrees.api.IResourcesTreesPlugin;
+import com.coolerpromc.resourcestrees.api.ResourcesTreesPlugin;
 import com.coolerpromc.resourcestrees.block.ModBlocks;
 import com.coolerpromc.resourcestrees.client.tint.ResourcesTypeTintSource;
 import com.coolerpromc.resourcestrees.platform.ForgeRegistryHelper;
@@ -14,17 +16,48 @@ import net.minecraft.world.level.block.ComposterBlock;
 import net.minecraftforge.event.AddPackFindersEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.listener.SubscribeEvent;
+import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.forgespi.language.ModFileScanData;
+import org.objectweb.asm.Type;
+
+import java.lang.reflect.InvocationTargetException;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 
 import static com.coolerpromc.resourcestrees.Constants.MODID;
 
 @Mod(Constants.MODID)
-public class ResourcesTrees {
-    public ResourcesTrees(FMLJavaModLoadingContext context) {
+public class ForgeResourcesTrees {
+    private static final Type TYPE = Type.getType(ResourcesTreesPlugin.class);
+    public static final List<IResourcesTreesPlugin> PLUGINS = new ArrayList<>();
+
+    public ForgeResourcesTrees(FMLJavaModLoadingContext context) {
+        List<ModFileScanData.AnnotationData> annotations = ModList.getAllScanData().stream().map(ModFileScanData::getAnnotations).flatMap(Collection::stream).filter(a -> TYPE.equals(a.annotationType())).toList();
+
+        for (ModFileScanData.AnnotationData annotationData : annotations){
+            String targetClassName = annotationData.memberName();
+
+            try {
+                Class<?> targetClass = Class.forName(targetClassName, false, IResourcesTreesPlugin.class.getClassLoader());
+
+                if (IResourcesTreesPlugin.class.isAssignableFrom(targetClass)) {
+                    IResourcesTreesPlugin plugin = (IResourcesTreesPlugin) targetClass.getDeclaredConstructor().newInstance();
+                    PLUGINS.add(plugin);
+                }
+
+            } catch (ClassNotFoundException e) {
+                e.printStackTrace();
+            } catch (InvocationTargetException | InstantiationException | IllegalAccessException | NoSuchMethodException e) {
+                throw new RuntimeException(e);
+            }
+        }
+
         ForgeRegistryHelper.register(context.getModBusGroup());
-        CommonClass.init();
+        ResourcesTrees.init();
 
         AddPackFindersEvent.BUS.addListener(this::onAddPackFinders);
         PlayerEvent.PlayerLoggedInEvent.BUS.addListener(this::onPlayerLoggedIn);